requirements.txt
discord.py
numpy
sympy
opencv-python
